homework_3/pyproject.toml
[project]
name = "homework-3-math-react-agent"
version = "0.1.0"
description = "ReAct agent for solving math and physics word problems using LangGraph + MCP + Wolfram Alpha"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Core agent dependencies
    "langchain-community>=0.3.24",
    "langchain-openai>=0.3.21",
    "langchain-google-genai>=2.0.8",
    "langgraph>=0.4.8",
    "litellm>=1.51.0",
    "python-dotenv>=1.1.0",

    # MCP server dependencies
    "mcp>=1.10.1",
    "starlette>=0.27.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "httpx>=0.28.1",
    "tavily-python>=0.5.0",

    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
